a68k: fix moveq #$FFFFFFFF,dn on 64-bit platforms. Bump version number, update history.
git-svn-id: svn+ssh://trac.godzil.net/gcc4ti@1365 9552661e-59e3-4036-b4f2-dbe53926924f

+++ trunk/external/a68k/source/Codegen.c
@@ -136,7 +136,7 @@ void GetObjectCode (__NO_PARAMS__) /* using now void in v.2.71.F3d - Kevin Kofle
     }
 
     /* added by Kevin Kofler in v.2.71.F3l: */
-    SrcVal=Src.Value;
+    SrcVal=(long)((int)(Src.Value));
     {struct AddrDiff *p=SrcAddrDiffs;
      while (p) {
        SrcVal+=p->paddr-p->naddr;
